今天的熬夜很没意义，快要被自己蠢哭了，一直不能运行的dao模式是因为自己忘了在configPlugin()配置mappingkit了，我日，顺便把Id从long改成int

+++ src/main/java/com/edulab/model_base/BaseUserAuths.java
@@ -23,8 +23,8 @@ public abstract class BaseUserAuths<M extends BaseUserAuths<M>> extends Model<M>
 		return (M)this;
 	}
 	
-	public java.lang.Long getUserId() {
-		return getLong("userId");
+	public java.lang.Integer getUserId() {
+		return getInt("userId");
 	}
 
 	public M setIdentityType(java.lang.String identityType) {
